pyproject.toml
[tool.poetry]
name = "fastapi-otel-elk"
version = "0.1.0"
description = "Fastapi with OpenTelemetry instrumentation"
readme = "README.md"
authors = ["Mátyás Sándor"]
packages = [
    {include = "app"}
]


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
opentelemetry-sdk = "^1.33.0"
opentelemetry-api = "^1.33.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.33.0"
opentelemetry-instrumentation-fastapi = "^0.54b0"
dynaconf = "^3.2.11"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
